Don't send a `ReadOk` unless message type is `read`
This fixes an issue that ultimately leads to Maelstrom crashing with an
`OutOfMemoryError` error during analysis; nodes are replying to
`broadcast_ok` messages, interpreting them as a `read` message. The
problem is likely related to Serde, as Serde can get confused by the
shape of the message when making a best attempt at deserializing message
data.

Post-fix, the number of interserver messages are reduced from roughly
2_000_000 to 1_000 with the following CLI options:

```
... --node-count 5 --time-limit 20 --rate 10
```

See `messages.svg` surfaced through the web UI for a visualization of
how messages are passed between nodes.

A long term fix would involve refactoring the response handler to only
invoke callbacks registered for the specific message type. Since it's
only happening to this message type, we can get away with taking a
shortcut for now.