Convert app to process messages asynchronously
This commit is a wholesale refactor to introduce the Tokio runtime for
async messages processing using async primitives. At this point, I'm
unsure any performance improvements are to be gained as a result, since
none of the processing involves blocking IO.

The refactor gives us the benefit of a slightly better app structure - a
clean entry point while removing `Node`'s responsiblity for outputting
results. With the addition of a few clarifying comments, we're setting
up the app for additional changes required by the challenge.

Some highlights:

* Use `Arc` and `Mutex` for mutating `Node` in a multi-threaded
  environment;
* Use channels for decoupling input/output from processing behavior;
* Updates to test to run async via `tokio::test`;

We're able to achieve the following results with this setup:

```
 :net {:all {:send-count 2536,
             :recv-count 2536,
             :msg-count 2536,
             :msgs-per-op 12.808081},
       :clients {:send-count 416, :recv-count 416, :msg-count 416},
       :servers {:send-count 2120,
                 :recv-count 2120,
                 :msg-count 2120,
                 :msgs-per-op 10.70707},
```

Compared to the Ruby demo `broadcast.rb`:

```
 :net {:all {:send-count 3806,
             :recv-count 3806,
             :msg-count 3806,
             :msgs-per-op 19.319798},
       :clients {:send-count 414, :recv-count 414, :msg-count 414},
       :servers {:send-count 3392,
                 :recv-count 3392,
                 :msg-count 3392,
                 :msgs-per-op 17.218273},
```

NOTE, the Ruby implementation ends up sending extra messages i.e. a
Node's non-neighbors, resulting in higher server messages sent and
therefore the appearance of higher throughput; check the `jepsen.log` in
the web UI for details.