Add `unacknowledged_messages` field to `Node`
The next exercise asks us to add a layer of durability to the Node's
broadcasting behavior. Maelstrom will introduce intermittent network
partitions which makes some nodes unreachable for a short period of
time. When this occurs, a Node must retry sending a message until it is
acknowledged by the recepient, avoiding message sending failure.

This is the first step where we're introducing message persistence,
considering that the Node's state must be mutably accessed across multiple
threads i.e. using an Arc + Mutex.