Tune timeout per broadcast message cycle
Sleeping for 1 millisecond is too short of a delay when a 100ms latency
is added to each message. My guess is with a shorter ms delay, the thread
resends more messages before the initial message is acknowledged,
compounded by every broadcast message the node receives. 1
second appears to be sufficient to allow the message to be
acknowledged before attempting to send a follow-up.

With this change, we've essentially completed challenges 3a - 3e (the
msg-per-op is still slightly high, clocking in at 23 msg/op) with
the caveat that the default `grid` topology is replaced with `tree4`:

```
./maelstrom test -w broadcast --bin ../tranquility/target/release/tranquility --node-count 25 --time-limit 20 --rate 100 --latency 100 --topology tree4
```

The final performance metrics and stable-latencies:

```
...
:stats {:valid? true,
       :count 1874,
       :ok-count 1874,
       :fail-count 0,
       :info-count 0,
       :by-f {:broadcast {:valid? true,
                          :count 919,
                          :ok-count 919,
                          :fail-count 0,
                          :info-count 0},
              :read {:valid? true,
                     :count 955,
                     :ok-count 955,
                     :fail-count 0,
                     :info-count 0}}},
:availability {:valid? true, :ok-fraction 1.0},
:net {:all {:send-count 47960,
           :recv-count 47960,
           :msg-count 47960,
           :msgs-per-op 25.592316},
     :clients {:send-count 3848, :recv-count 3848, :msg-count 3848},
     :servers {:send-count 44112,
               :recv-count 44112,
               :msg-count 44112,
               :msgs-per-op 23.538954},
     :valid? true},
...

```

```
:stable-latencies {0 0,
                   0.5 373,
                   0.95 488,
                   0.99 502,
                   1 517},
```